Implement author_titles_validator
Implement author_titles_validator for the fuzzy matcher with tests and
needed utils. It computes a validation score for the possible match. The
score is based on the Jaccard index of the authors sets and the maximum
Jaccard index found between 2 titles from the record and the possible
match title sets. The default score is 0.5. If the computed score is
higher than this, then the match is valid, otherwise it is not.

Sem-Ver: new feature
Signed-off-by: Iuliana Voinea <iuliana.voinea@student.manchester.ac.uk>